+++ extensions/quick-access/CHANGELOG.md
@@ -1,9 +1,10 @@
 # Quick Access Changelog
 
-## [Add Open With] - 2022-04-25
+## [Add Open With] - 2022-04-26
 
-- Add actions "Open With..."
+- Add actions: Open With...
 - Update shortcuts for certain actions
+- Add preferences: Primary action
 
 ## [Initial Version] - 2022-04-22